requirements.txt
Kivy==2.3.1
Kivy-Garden==0.1.5
open3d>=0.18.0
opencv-python==4.11.0.86
pydantic==2.10.6
pydantic_core==2.27.2
scikit-image>=0.24.0
scipy>=1.13.0
torch==2.6.0
torchvision>=0.21.0
pygrabber
Pillow>=10.0.0
tqdm>=4.65.0

